+++ spec/features/user_can_submit_posts_spec.rb
@@ -1,6 +1,16 @@
 require 'rails_helper'
 
 RSpec.feature "Timeline", type: :feature do
+  scenario 'User can submit posts from the homepage' do
+    sign_up
+
+    visit('/posts')
+    fill_in "new-post-message", with: 'This is a post'
+    click_button "Post"
+
+    expect(page).to have_content("This is a post")
+  end
+  
   scenario "Can submit posts and view them" do
     sign_up